Initial Commit
This program does a basic function. It asks you to input the total amount of the bill, the percentage of tip you would like to give, and how many people are splitting the bill. Once these three variables have values, the program calculates your tip and total amount each person should pay.

Later updates of this program will prevent the program from breaking when a value other than an integer or float are added to the program.