Initial Commit
This program is an expanded version of a game based on a small project from a beginner Python tutorial. There are a few bugs that I am trying to fix. The game starts as desired, but rather than running in an infinite loop (as desired), it stops after 2 rolls and then requires a restart.

Another notable bug is that upon starting the game, if you choose the N(No) option when asked to roll, the program still moves forward with rolling anyway. If you select the N(No) option on the second roll, the game ends (program exits) as intended.