Initial Commit
The else statement isn't behaving the way it is supposed to. Print is added to keep the code from breaking if a size other than S M or L is chosen. The while loop is intended to allow the code to keep running until the appropriate selection has been made.

Other planned adds are making the inputs non-case sensitive to make this program more user friendly.