Initial Commit
The instructions were to write a program that finds the highest score in a list of inputted numbers without using the max() function. I was able to accomplish this using a for loop, an if statement, and creating two variables with value 0. I used the for loop and if statement to assign new values to the variables. See the Pseudocode for logic used in this program. Note: This program will break if anything other than integers are input. For the purpose of this basic assignment, I am not adding any additional code to fix this program.